# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "micro de seguridad.",
        "title": "medfri-security",
        "contact": {
            "name": "Soporte de API",
            "url": "http://www.soporte-api.com",
            "email": "soporte@api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/medfri-security",
    "paths": {
...
        "/entity": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite actualizar la información de una entidad existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entidades"
                ],
                "summary": "Actualizar una entidad",
                "parameters": [
                    {
                        "description": "Información de la entidad actualizada",
                        "name": "entity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Entity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entidad actualizada con éxito",
                        "schema": {
                            "$ref": "#/definitions/entity.Entity"
                        }
                    },
                    "400": {
                        "description": "Error en el cuerpo de la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite crear una nueva entidad en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entidades"
                ],
                "summary": "Crear una entidad",
                "parameters": [
                    {
                        "description": "Información de la entidad",
                        "name": "entity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Entity"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Entidad creada con éxito",
                        "schema": {
                            "$ref": "#/definitions/entity.Entity"
                        }
                    },
                    "400": {
                        "description": "Error en el cuerpo de la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/entity/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint obtener todas las entidades",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entidades"
                ],
                "summary": "Obtener todas las entidades",
                "responses": {
                    "200": {
                        "description": "Entidades encontrada",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Entity"
                            }
                        }
                    },
                    "404": {
                        "description": "Entidades no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/entity/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite obtener la información de una entidad específica usando su ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entidades"
                ],
                "summary": "Obtener una entidad por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la entidad",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entidad encontrada",
                        "schema": {
                            "$ref": "#/definitions/entity.Entity"
                        }
                    },
                    "404": {
                        "description": "Entidad no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite eliminar una entidad específica usando su ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entidades"
                ],
                "summary": "Eliminar una entidad",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la entidad",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Entidad eliminada con éxito"
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/rol-recurso": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite actualizar la información de un rol recurso existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles-recursos"
                ],
                "summary": "Actualizar un rol recurso",
                "parameters": [
                    {
                        "description": "Información del rol recurso actualizada",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RoleResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "rol recurso actualizado con éxito",
                        "schema": {
                            "$ref": "#/definitions/entity.RoleResource"
                        }
                    },
                    "400": {
                        "description": "Error en el cuerpo de la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rol-recurso/asignar": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite crear un nuevo rol recurso en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles-recursos"
                ],
                "summary": "Crear un rol recurso",
                "parameters": [
                    {
                        "description": "Información del rol recurso",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RoleResource"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Rol recurso creado con éxito",
                        "schema": {
                            "$ref": "#/definitions/entity.RoleResource"
                        }
                    },
                    "400": {
                        "description": "Error en el cuerpo de la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rol-recurso/desasignar/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint permite eliminar un rol recurso existente en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles-recursos"
                ],
                "summary": "Eliminar un rol recurso",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del rol recurso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol recurso eliminado con éxito",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rol-recurso/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Este endpoint devuelve la información de un rol recurso específico dado su ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles-recursos"
                ],
                "summary": "Obtener un rol recurso por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del rol recurso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "rol recurso encontrado",
                        "schema": {
                            "$ref": "#/definitions/entity.RoleResource"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Entity": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "entidad_id": {
                    "type": "integer"
                },
                "nit": {
                    "type": "integer"
                },
                "nit_alterno": {
                    "type": "string"
                },
                "razon_social": {
                    "type": "string"
                }
            }
        },
...
        "entity.Resource": {
            "type": "object",
            "properties": {
                "acceso": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "recurso_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Rol": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string"
                },
                "entidad_id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "rolID": {
                    "type": "integer"
                }
            }
        },
        "entity.RoleResource": {
            "type": "object",
            "properties": {
                "acceso": {
                    "type": "string"
                },
                "entidad": {
                    "$ref": "#/definitions/entity.Entity"
                },
                "entidad_id": {
                    "type": "integer"
                },
                "recurso": {
                    "$ref": "#/definitions/entity.Resource"
                },
                "recurso_id": {
                    "type": "integer"
                },
                "rol": {
                    "$ref": "#/definitions/entity.Rol"
                },
                "rolRecursoID": {
                    "type": "integer"
                },
                "rol_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ingresa \"Bearer {token}\" para autenticar.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}